requirements/_base.txt
Django~=5.2
djangorestframework~=3.15
psycopg2-binary~=2.9
python-decouple~=3.8